requirements.txt
altair==6.0.0 ; python_version == "3.13"
annotated-types==0.7.0 ; python_version == "3.13"
anyio==4.13.0 ; python_version == "3.13"
attrs==26.1.0 ; python_version == "3.13"
beautifulsoup4==4.14.3 ; python_version == "3.13"
blinker==1.9.0 ; python_version == "3.13"
cachetools==7.0.5 ; python_version == "3.13"
certifi==2026.2.25 ; python_version == "3.13"
charset-normalizer==3.4.6 ; python_version == "3.13"
click==8.3.1 ; python_version == "3.13"
colorama==0.4.6 ; python_version == "3.13" and platform_system == "Windows"
distro==1.9.0 ; python_version == "3.13"
gitdb==4.0.12 ; python_version == "3.13"
gitpython==3.1.46 ; python_version == "3.13"
h11==0.16.0 ; python_version == "3.13"
httpcore==1.0.9 ; python_version == "3.13"
httpx==0.28.1 ; python_version == "3.13"
idna==3.11 ; python_version == "3.13"
jinja2==3.1.6 ; python_version == "3.13"
jiter==0.13.0 ; python_version == "3.13"
jsonschema-specifications==2025.9.1 ; python_version == "3.13"
jsonschema==4.26.0 ; python_version == "3.13"
markupsafe==3.0.3 ; python_version == "3.13"
narwhals==2.18.1 ; python_version == "3.13"
numpy==2.4.3 ; python_version == "3.13"
openai==2.30.0 ; python_version == "3.13"
packaging==26.0 ; python_version == "3.13"
pandas==2.3.3 ; python_version == "3.13"
pillow==12.1.1 ; python_version == "3.13"
protobuf==6.33.6 ; python_version == "3.13"
psycopg-binary==3.3.3 ; python_version == "3.13" and implementation_name != "pypy"
psycopg==3.3.3 ; python_version == "3.13"
pyarrow==23.0.1 ; python_version == "3.13"
pydantic-core==2.41.5 ; python_version == "3.13"
pydantic==2.12.5 ; python_version == "3.13"
pydeck==0.9.1 ; python_version == "3.13"
python-dateutil==2.9.0.post0 ; python_version == "3.13"
python-dotenv==1.2.2 ; python_version == "3.13"
pytz==2026.1.post1 ; python_version == "3.13"
referencing==0.37.0 ; python_version == "3.13"
requests==2.33.0 ; python_version == "3.13"
rpds-py==0.30.0 ; python_version == "3.13"
six==1.17.0 ; python_version == "3.13"
smmap==5.0.3 ; python_version == "3.13"
sniffio==1.3.1 ; python_version == "3.13"
soupsieve==2.8.3 ; python_version == "3.13"
streamlit==1.55.0 ; python_version == "3.13"
tenacity==9.1.4 ; python_version == "3.13"
toml==0.10.2 ; python_version == "3.13"
tornado==6.5.5 ; python_version == "3.13"
tqdm==4.67.3 ; python_version == "3.13"
typing-extensions==4.15.0 ; python_version == "3.13"
typing-inspection==0.4.2 ; python_version == "3.13"
tzdata==2025.3 ; python_version == "3.13"
urllib3==2.6.3 ; python_version == "3.13"
watchdog==6.0.0 ; python_version == "3.13" and platform_system != "Darwin"
